Add class distance to null rejection coefficient translator.
This commit adds classDistanceToNullRejectionCoefficient(), a function
that converts (label-specific) class distances into the units of the
null rejection coefficients. That is, it does the inverse of the
calculation that converts null rejection coefficients to null rejection
thresholds.

It also adds a getSupportsClassDistanceToNullRejectionCoefficient(),
which indicates whether or not the given classifier actually supports
the classDistanceToNullRejectionCoefficient() — i.e. overrides it with
a proper implementation.